- 新增 U 币支付相关控制器、服务、设置页与充值弹窗流程
- 补充 Antom、品牌图标、多语言文案和支付方式展示逻辑
- 合并用户模型字段冲突，保留分销商模板、U 币地址与供应商字段

+++ controller/topup_waffo.go
@@ -168,9 +168,9 @@ func RequestWaffoPay(c *gin.Context) {
 	paymentRequestId := merchantOrderId
 
 	// Token 模式下归一化 Amount（存等价美元/CNY 数量，避免 RechargeWaffo 双重放大）
-	amount := req.Amount
+	amount := float64(req.Amount)
 	if operation_setting.GetQuotaDisplayType() == operation_setting.QuotaDisplayTypeTokens {
-		amount = int64(float64(req.Amount) / common.QuotaPerUnit)
+		amount = float64(req.Amount) / common.QuotaPerUnit
 		if amount < 1 {
 			amount = 1
 		}
